+++ examples/using_bag_base.py
@@ -11,31 +11,33 @@ ch_formatter = logging.Formatter('%(levelname)-9s %(name)s.%(funcName)s:%(lineno
 ch.setFormatter(ch_formatter)
 logger.addHandler(ch)
 
+# Example that use bag.base to:
+# - test for being a BAG file
+# - open and read some metadata info from a BAG file
 
 from hydroffice.bag.base import is_bag, File
-from hydroffice.bag.helper import Helper, BAGError
-
+from hydroffice.bag.helper import Helper
 
 file_bag_0 = os.path.join(Helper.samples_folder(), "bdb_00.bag")
 if os.path.exists(file_bag_0):
-    print("- file_bag_0: %s is BAG? %r" % (file_bag_0, is_bag(file_bag_0)))
+    logger.info("- file_bag_0: %s is BAG? %r" % (file_bag_0, is_bag(file_bag_0)))
 
 file_bag_1 = os.path.join(Helper.samples_folder(), "bdb_01.bag")
 if os.path.exists(file_bag_1):
-    print("- file_bag_1: %s is BAG? %r" % (file_bag_1, is_bag(file_bag_1)))
+    logger.info("- file_bag_1: %s is BAG? %r" % (file_bag_1, is_bag(file_bag_1)))
 
 file_fake_0 = os.path.join(Helper.samples_folder(), "fake_00.bag")
 if os.path.exists(file_fake_0):
-    print("- file_fake_0: %s is BAG? %r" % (file_fake_0, is_bag(file_fake_0)))
+    logger.info("- file_fake_0: %s is BAG? %r" % (file_fake_0, is_bag(file_fake_0)))
 
 bag_0 = File(file_bag_0)
-print(bag_0)
-print("filename: %s" % bag_0.filename)
-print("attributes for %s: %d" % (bag_0.attrs, len(bag_0.attrs)))
-print("driver: %s" % bag_0.driver)
+logger.info("\n%s\n" % bag_0)
+logger.info("filename: %s" % bag_0.filename)
+logger.info("attributes for %s: %d" % (bag_0.attrs, len(bag_0.attrs)))
+logger.info("driver: %s" % bag_0.driver)
 bag_0.flush()
 bag_0.close()
-print(bag_0)
+logger.info("\n%s\n" % bag_0)
 
 bag_1 = File(file_bag_1)
 bag_1.flush()
@@ -44,7 +46,7 @@ bag_1.close()
 try:
     fake_0 = File(file_fake_0)
 except IOError:
-    print("Expected exception")
+    logger.info("Expected exception")